backend/requirements.txt
alembic==1.13.1
celery==5.3.6
fastapi==0.110.0
httpx==0.27.0
passlib[bcrypt]==1.7.4
prometheus-client==0.20.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
python-json-logger==2.0.7
redis==5.0.4
slowapi==0.1.9
SQLAlchemy==2.0.29
starlette==0.37.2
uvicorn[standard]==0.29.0

# Observability
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-distro==0.47b0

# Dev / QA
pytest==8.2.0
ruff==0.4.5
mypy==1.10.0